create frame comparison module
